MP-187 :truck: Rename: modules.jwt 내부 폴더 구조 통일화
* 우리 프로젝트에서의 표준 폴더 구조와 동일하게 함
