+++ src/main/java/kr/modusplant/modules/jwt/app/service/TokenProvider.java
@@ -2,8 +2,8 @@ package kr.modusplant.modules.jwt.app.service;
 
 import io.jsonwebtoken.*;
 import jakarta.annotation.PostConstruct;
-import kr.modusplant.modules.jwt.app.error.InvalidTokenException;
-import kr.modusplant.modules.jwt.app.error.TokenKeyCreationException;
+import kr.modusplant.modules.jwt.error.InvalidTokenException;
+import kr.modusplant.modules.jwt.error.TokenKeyCreationException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
