+++ src/main/java/kr/modusplant/modules/jwt/error/TokenKeyCreationException.java
@@ -1,4 +1,4 @@
-package kr.modusplant.modules.jwt.app.error;
+package kr.modusplant.modules.jwt.error;
 
 public class TokenKeyCreationException extends AuthTokenException {
     public TokenKeyCreationException(String message, Throwable cause) {
